Separate out is launching functionality
